refactor: decompose god-components — pixel-blast, regex hooks, explain, dino [CU-07,CB-04,LR-01,CG-04]
Phase 10 (wave A, behavior-preserving decomposition):
- CU-07 lib/animations/pixel-blast-engine.ts: extract GLSL (FRAGMENT_SRC/VERTEX_SRC),
  SHAPE_MAP, createTouchTexture, createLiquidEffect (no React). pixel-blast.tsx
  746 -> 435 LOC; effect body kept intact (calls engine fns) to preserve disposal/
  RAF/listener lifecycle. Visual output identical.
- CB-04 hooks/use-regex-url-state.ts + use-regex-shortcuts.ts: extract URL-hash
  hydration/write-back and the keyboard-shortcut hub from RegexTesterClient
  (298 -> 265 LOC). Exact key-combo + debounce behavior preserved.
- LR-01 lib/regex-tester/explain.ts: table-drive the 8 open-arms via OPEN_META +
  unify the 2 close-arms into closeFrame(); all 14 explain tests pass unchanged.
- CG-04 dino split into dino-original.tsx (iframe) + dino-custom-game.tsx (canvas)
  + thin dino-client dispatcher; lift leaked dims (canvasHeight, obstacleW/H,
  groundTileWidth) into GAME_CONFIG. Gameplay identical.

Verify: 408 tests pass, tsc clean, build green, lint unchanged.
